Remove duplicate tests and combine quote test
